--- a/nltk/metrics/distance.py
+++ b/nltk/metrics/distance.py
@@ -22,8 +22,6 @@ As metrics, they must satisfy the following three requirements:
 import operator
 import warnings
 
-import stringzilla as sz
-
 
 def _edit_dist_init(len1, len2):
     lev = []
@@ -91,7 +89,9 @@ def edit_distance(s1, s2, substitution_cost=1, transpositions=False):
     # StringZilla currently doesn't support transpositions and cost-parameterized substitutions.
     # It's C implementation is 10x faster than a pure Python implementation.
     if not transpositions and substitution_cost == 1:
-        return sz.edit_distance_unicode(s1, s2)
+        from stringzilla import edit_distance_unicode as sz_edit_dist
+
+        return sz_edit_dist(s1, s2)
 
     # set up a 2-D array
     len1 = len(s1)