--- a/pip-req.txt
+++ b/pip-req.txt
@@ -13,3 +13,4 @@ click>=7.1.2
 joblib>=1.0.1
 tqdm>=4.59.0
 pre-commit>=2.13.0
+stringzilla>=3.1.2