This PR replaces the native Python implementation of the Levenshtein distance with a SIMD-accelerated version from the StringZilla library.

On ~5 letter words from a typical English corpus - StringZilla is over 10x faster than current Python implementation in NLTK (1.24 s ± 57.4 ms vs 13.5 s ± 287 ms per loop across 7 runs). On a multilingual corpus with longer words - same result. 

1. StringZilla uses less `if` branches than most native implementations, so it's faster than many implementations in the serial mode. When AVX-512 is available it can use specialized assembly instructions to accelerate evaluation.
2. StringZilla currently [provides wheels](https://pypi.org/project/stringzilla/#files) for {manylinux, musllinux, macos, windows} × { 32-bit, 64-bit } × { x86, Arm, PowerPC } × [ Python 3.6, Python 3.12 ] making it practically the most portable hardware-accelerated library in the Python ecosystem.