@alvations 2024-07-08
While this is a good speedup, it adds additional dependencies. I'll avoid it and users looking for speedier alternatives should be able to find `StringZilla` or other.