@ekaf 2024-12-02
I would prefer an open source implementation in Python, allowing students to learn the classical algorithms. Compiled code is the contrary of what I would like to see in NLTK.